Move lambda logic to partial class
